+++ source/win-main/GlobalSearch.vue
@@ -25,8 +25,14 @@
       <ButtonControl
         v-bind:label="searchButtonLabel"
         v-bind:inline="true"
-        v-bind:disabled="filesToSearch.length > 0"
+        v-bind:disabled="false"
         v-on:click="startSearch()"
+      ></ButtonControl>
+      <ButtonControl
+        v-bind:label="cancelButtonLabel"
+        v-bind:inline="true"
+        v-bind:disabled="!searchIsRunning"
+        v-on:click="cancelSearch()"
       ></ButtonControl>
     </p>
     <!-- ... as well as two buttons to clear the results or toggle them. -->
@@ -61,7 +67,7 @@
           v-bind:max="sumFilesToSearch"
           v-bind:value="sumFilesToSearch - filesToSearch.length"
           v-bind:interruptible="true"
-          v-on:interrupt="filesToSearch = []"
+          v-on:interrupt="cancelSearch()"
         ></ProgressControl>
       </div>
       <hr>
@@ -161,6 +167,7 @@ const filterLabel = trans('Filter search results')
 const restrictDirLabel = trans('Restrict search to directory')
 const restrictDirPlaceholder = trans('Choose directory…')
 const searchButtonLabel = trans('Search')
+const cancelButtonLabel = trans('Cancel')
 const clearButtonLabel = trans('Clear search')
 const toggleButtonLabel = trans('Toggle results')
 
@@ -268,6 +275,9 @@ const filteredSearchResults = computed<SearchResultWrapper[]>(() => {
   })
 })
 
+const searchIsRunning = computed(() => { return filesToSearch.value.length > 0 })
+const shouldStartNewSearch = ref<boolean>(false)
+
 watch(fileTree, () => {
   recomputeDirectorySuggestions()
 })
@@ -297,14 +307,14 @@ function recomputeDirectorySuggestions (): void {
 }
 
 function startSearch (overrideQuery?: string): void {
-  if (filesToSearch.value.length > 0) {
-    console.warn('Global search in progress: Not starting a new one.')
-    return
-  }
-
   // This allows other components to inject a new query when starting a search
   if (overrideQuery !== undefined) {
     query.value = overrideQuery
+  }
+
+  if (searchIsRunning.value) {
+    cancelSearch(true)
+    return
   }
 
   // We should start a search. We need two types of information for that:
@@ -429,8 +439,17 @@ async function singleSearchRun (): Promise<void> {
   finaliseSearch()
 }
 
+function cancelSearch (startNewSearch: boolean = false): void {
+  filesToSearch.value = []
+  shouldStartNewSearch.value = startNewSearch
+}
+
 function finaliseSearch (): void {
   filesToSearch.value = [] // Reset, in case the search was aborted.
+  if (shouldStartNewSearch.value) {
+    shouldStartNewSearch.value = false
+    startSearch()
+  }
 }
 
 function emptySearchResults (): void {